test: split sdf materialize default-resolution test into its own file for CI shard balance
materialize.test.ts ran 62s as one file; the default-resolution
(res=30) materialize is the slowest case (~29s) and moves to
materialize.defaultRes.test.ts with its own OCCT init. No dedup
possible: every test materializes a different field/resolution. Test
counts unchanged: 5 -> 4+1. Measured per-file: 8.4s + 16.3s.
